Update patch set 1
Patch Set 1:

The version of Perl (5.22.1) included in Ubuntu 16.04 doesn't work with the old Perl script. Those of us on 14.04, Perl 5.14.2 have no problems. This will become important in the future when the old Perl version becomes less common. I submitted it because of chatter about this problem that I have heard concerning other devices. This is a simple fix.

Patch-set: 1